Removes DockerExtension and improves exception message for Docker env parsing.
